requirements.txt
streamlit==1.25.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==0.27.2
